+++ src/ProSuite.QA.Tests/Transformers/TrDissolve.cs
@@ -194,12 +194,12 @@ namespace ProSuite.QA.Tests.Transformers
 
 			public override IEnvelope Extent => _dissolve.Extent;
 
-			public override VirtualRow GetUncachedRow(int id)
+			public override VirtualRow GetUncachedRow(long id)
 			{
 				throw new NotImplementedException();
 			}
 
-			public override int GetRowCount(IQueryFilter queryFilter)
+			public override long GetRowCount(IQueryFilter queryFilter)
 			{
 				return Search(queryFilter, true).Count();
 				// TODO: Consider new Method GetRowCountEstimate()? Or add progress token to Search()?
