+++ src/ProSuite.QA.Tests/Transformers/TrGeometryTransform.cs
@@ -136,7 +136,7 @@ namespace ProSuite.QA.Tests.Transformers
 
 			public IGeometryTransformer Transformer { get; }
 
-			public override GdbRow CreateObject(int oid,
+			public override GdbRow CreateObject(long oid,
 			                                    IValueList valueList = null)
 			{
 				var joinedValueList = new MultiListValues();
@@ -190,7 +190,7 @@ namespace ProSuite.QA.Tests.Transformers
 
 		private class TransformedFeature : GdbFeature
 		{
-			public TransformedFeature(int oid, TransformedFc featureClass,
+			public TransformedFeature(long oid, TransformedFc featureClass,
 			                          MultiListValues valueList)
 				: base(oid, featureClass, valueList) { }
 
@@ -219,12 +219,12 @@ namespace ProSuite.QA.Tests.Transformers
 
 			public override IEnvelope Extent => _t0.Extent;
 
-			public override VirtualRow GetUncachedRow(int id)
+			public override VirtualRow GetUncachedRow(long id)
 			{
 				throw new NotImplementedException();
 			}
 
-			public override int GetRowCount(IQueryFilter queryFilter)
+			public override long GetRowCount(IQueryFilter queryFilter)
 			{
 				// TODO
 				return _t0.RowCount(queryFilter);
